+++ test/unit/argparse_test.lua
@@ -76,14 +76,6 @@ g.test_sections = function()
             ['custom.sub.sub'] = {
                 x = '@custom.sub.sub',
                 y_subsub = true,
-            },
-            ['custom.x.y.z'] = {
-                x = '@custom.x.y.z',
-                y_xyz = true,
-            },
-            ['myapp.instance'] = {
-                x = '@myapp.instance',
-                y_instance = true,
             }
         })
     )
@@ -112,10 +104,11 @@ g.test_sections = function()
         y_default = 0,
     })
 
-    t.assert_error_msg_contains(
-        'ParseConfigError: Missing section: unknown',
-        check, '--instance-name unknown'
-    )
+    check('--instance-name unknown', {
+        instance_name = 'unknown',
+        x = '@default',
+        y_default = 0,
+    })
 
     check('--instance-name custom', {
         instance_name = 'custom',
@@ -124,10 +117,12 @@ g.test_sections = function()
         y_custom = '$',
     })
 
-    t.assert_error_msg_contains(
-        'ParseConfigError: Missing section: custom.bad',
-        check, '--instance-name custom.bad'
-    )
+    check('--instance-name custom.bad', {
+        instance_name = 'custom.bad',
+        x = '@custom',
+        y_default = 0,
+        y_custom = '$',
+    })
 
     check('--instance-name custom.sub', {
         instance_name = 'custom.sub',
@@ -135,20 +130,6 @@ g.test_sections = function()
         y_default = 0,
         y_custom = '$',
         y_sub = 3.14,
-    })
-
-    check('--app-name myapp', {
-        app_name = 'myapp',
-        x = '@default',
-        y_default = 0,
-    })
-
-    check('--app-name myapp --instance-name instance', {
-        app_name = 'myapp',
-        instance_name = 'instance',
-        x = '@myapp.instance',
-        y_default = 0,
-        y_instance = true,
     })
 
     check('--instance_name custom.sub.sub', {
@@ -159,11 +140,6 @@ g.test_sections = function()
         y_sub = 3.14,
         y_subsub = true,
     })
-
-    t.assert_error_msg_contains(
-        'ParseConfigError: Missing section: custom.x\n',
-        check, '--instance-name custom.x.y.z'
-    )
 end
 
 g.test_priority = function()
@@ -196,6 +172,7 @@ g.test_priority = function()
 
     check('--instance-name custom',             {'TARANTOOL_CFG=./x.yml'}, '@custom')
     check('', {'TARANTOOL_INSTANCE_NAME=custom', 'TARANTOOL_CFG=./x.yml'}, '@custom')
+    check('', {'TARANTOOL_INSTANCE_NAME=CUSTOM', 'TARANTOOL_CFG=./x.yml'}, '@default')
 
     check('--cfg ./cfg.yml',   {'TARANTOOL_CFG=./x.yml'}, '@cfg-default')
     check('',              {'TARANTOOL_CFG="./cfg.yml"'}, '@cfg-default')
@@ -320,17 +297,6 @@ g.test_badfile = function()
     )
     local ret = g.run('--cfg tarantool.yml', {}, {ignore_errors = true})
     t.assert_str_contains(ret.err, 'DecodeYamlError: tarantool.yml: unexpected END event')
-
-    utils.file_write(
-        fio.pathjoin(g.tempdir, 'tarantool.yml'),
-        yaml.encode({['app_name.instance_name'] = {x = 'sup'}})
-    )
-    local ret = g.run(
-        '--cfg tarantool.yml' ..
-        ' --app_name app_name' ..
-        ' --instance_name harry',
-        {}, {ignore_errors = true})
-    t.assert_str_contains(ret.err, 'ParseConfigError: Missing section: app_name.harry')
 end
 
 g.test_box_opts = function()